Improve day07 and dynamic arrays

+++ dynamic_arrays/src/ft_dynamic_size_t.c
@@ -1,18 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_dynamic_int.c                                   :+:      :+:    :+:   */
+/*   ft_dynamic_size_t.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: rselva-2 <rselva-2@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/11/27 17:06:30 by rselva-2          #+#    #+#             */
-/*   Updated: 2025/12/07 22:28:46 by rselva-2         ###   ########.fr       */
+/*   Created: 2025/12/07 22:34:40 by rselva-2          #+#    #+#             */
+/*   Updated: 2025/12/07 22:34:42 by rselva-2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_dynarray.h"
 
-int	add_int(t_dyn_int *d_array, size_t index, int number)
+int	add_int(t_dyn_int *d_array, size_t index, size_t number)
 {
 	size_t	new_size;
 
@@ -22,7 +22,7 @@ int	add_int(t_dyn_int *d_array, size_t index, int number)
 	if (new_size != d_array->size)
 	{
 		if (ft_realloc((void **)&(d_array->arr),
-			d_array->size * sizeof(int), new_size * sizeof(int)))
+			d_array->size * sizeof(size_t), new_size * sizeof(size_t)))
 			d_array->size = new_size;
 		else
 			return (0);
@@ -34,7 +34,7 @@ int	add_int(t_dyn_int *d_array, size_t index, int number)
 int	init_dyn_int(t_dyn_int *d_array, size_t size)
 {
 	d_array->size = size;
-	d_array->arr = malloc(d_array->size * sizeof(int));
+	d_array->arr = malloc(d_array->size * sizeof(size_t));
 	if (!(d_array->arr))
 	{
 		d_array->size = 0;
